requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
mysqlclient==2.1.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
